Rename "colour.algebra.random" module to "colour.algebra.prng" to prevent import exceptions.
