+++ colour/algebra/prng.py
